backend/requirements.txt
# Backend requirements
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
werkzeug==2.3.7
python-dotenv==1.0.0
pillow==10.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.7
bcrypt==4.0.1
marshmallow==3.20.1
requests==2.31.0